Move JWS verification code to verify.go
This commit moves `signatureValidationError`, `extractJWSKey` and
`verifyPOST` from `wfe2/wfe.go` to `wfe2/verify.go`. This is in
preparation of refactoring for the ACME v2 POST verification logic to
help keep diffs reviewable.
